// green-store/app/lib/api/models/order-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Store Management API
 * API for managing store operations including inventory, orders, and customers
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { OrderItemResponse } from './order-item-response';

export interface OrderResponse {
    'id'?: number;
    'customerName'?: string | null;
    'employeeName'?: string | null;
    'totalAmount'?: number;
    'discountAmount'?: number;
    'status'?: string | null;
    'orderDate'?: string;
    'listItem'?: Array<OrderItemResponse> | null;
}