// green-store/app/lib/api/models/order-item-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Store Management API
 * API for managing store operations including inventory, orders, and customers
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface OrderItemResponse {
    'id'?: number;
    'productName'?: string | null;
    'quantity'?: number;
    'price'?: number;
    'subTotal'?: number;
}